Move Reader class to the correct namespace
