+++ SAGESharp.Tests/BKDSerializationTests.cs
@@ -6,6 +6,7 @@
 using FluentAssertions;
 using NUnit.Framework;
 using SAGESharp.IO;
+using SAGESharp.IO.Binary;
 using System.Collections.Generic;
 using System.IO;
