+++ SAGESharp/BKD.cs
@@ -6,6 +6,7 @@
 using Equ;
 using NUtils.Extensions;
 using SAGESharp.IO;
+using SAGESharp.IO.Binary;
 using System;
 using System.Collections.Generic;
